requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pyttsx3
pytz
SpeechRecognition
langdetect
google-generativeai
streamlit
